refactor(crypto): improve discoverability of the package by further separating classes and functions
